113 commit: Добавил функционал для работы исполнетелей на этапах проекта.

+++ models/RespondsMvp.php
@@ -5,6 +5,7 @@ namespace app\models;
 use app\models\interfaces\RespondsInterface;
 use app\models\traits\SoftDeleteModelTrait;
 use Throwable;
+use Yii;
 use yii\base\ErrorException;
 use yii\db\ActiveQuery;
 use yii\db\ActiveRecord;
@@ -25,6 +26,8 @@ use yii\helpers\FileHelper;
  * @property int $date_plan                         Плановая дата интервью
  * @property string $place_interview                Место проведения интервью
  * @property int|null $deleted_at                   Дата удаления
+ * @property int|null $contractor_id                Идентификатор исполнителя, который опросил респондента (если null - опрос проводил проектант)
+ * @property int|null $task_id                      Идентификатор задания исполнителя, по которому исполнитель опросил респондента (если null - опрос проводил проектант)
  *
  * @property ConfirmMvp $confirm                    Подтверждение Mvp-продукта
  * @property InterviewConfirmMvp $interview         Информация о проведении интервью
@@ -96,6 +99,7 @@ class RespondsMvp extends ActiveRecord implements RespondsInterface
             [['name'], 'string', 'max' => 100],
             [['info_respond', 'email', 'place_interview'], 'string', 'max' => 255],
             ['email', 'email', 'message' => 'Неверный формат адреса электронной почты'],
+            [['contractor_id', 'task_id'], 'safe'],
         ];
     }
 
@@ -136,10 +140,11 @@ class RespondsMvp extends ActiveRecord implements RespondsInterface
     }
 
     /**
+     * @param bool $isRemoveAnswers
      * @return void
      * @throws ErrorException
      */
-    private function deleteDataRespond(): void
+    private function deleteDataRespond(bool $isRemoveAnswers = true): void
     {
         $confirm = ConfirmMvp::findOne($this->getConfirmId());
         $mvp = Mvps::findOne($confirm->getMvpId());
@@ -155,7 +160,11 @@ class RespondsMvp extends ActiveRecord implements RespondsInterface
         }
         //Удаление ответов респондента на вопросы интервью
         if (AnswersQuestionsConfirmMvp::findAll(['respond_id' => $this->getId()])) {
-            AnswersQuestionsConfirmMvp::deleteAll(['respond_id' => $this->getId()]);
+            if ($isRemoveAnswers) {
+                AnswersQuestionsConfirmMvp::deleteAll(['respond_id' => $this->getId()]);
+            } else {
+                AnswersQuestionsConfirmMvp::updateAll(['answer' => ''], ['respond_id' => $this->getId()]);
+            }
         }
         //Удаление дирректории респондента
         $del_dir = UPLOAD.'/user-'.$user->getId().'/project-'.$project->getId().'/segments/segment-'.$segment->getId().'/problems/problem-'.$problem->getId().
@@ -170,6 +179,43 @@ class RespondsMvp extends ActiveRecord implements RespondsInterface
             FileHelper::removeDirectory($cachePathDelete);
         }
     }
+
+
+    /**
+     * @return bool
+     */
+    public function clearData(): bool
+    {
+        $transaction = Yii::$app->db->beginTransaction();
+
+        try {
+            $responds = self::findAll(['confirm_id' => $this->getConfirmId()]);
+            $numberResponds = [];
+            foreach ($responds as $respond) {
+                if (preg_match('/^Респондент \d+$/', $respond->getName())) {
+                    $numberResponds[] = str_replace('Респондент ', '', $respond->getName());
+                }
+            }
+
+            $this->name = 'Респондент ' . (max($numberResponds) + 1);
+            $this->info_respond = '';
+            $this->email = '';
+            $this->date_plan = null;
+            $this->place_interview = '';
+            $this->contractor_id = null;
+            $this->task_id = null;
+            $this->save();
+            $this->deleteDataRespond(false);
+
+            $transaction->commit();
+            return true;
+
+        } catch (\Exception $exception) {
+            $transaction->rollBack();
+            return false;
+        }
+    }
+
 
     /**
      * @return int
@@ -289,5 +335,37 @@ class RespondsMvp extends ActiveRecord implements RespondsInterface
     public function setDeletedAt(int $deleted_at): void
     {
         $this->deleted_at = $deleted_at;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getContractorId(): ?int
+    {
+        return $this->contractor_id;
+    }
+
+    /**
+     * @param int $contractor_id
+     */
+    public function setContractorId(int $contractor_id): void
+    {
+        $this->contractor_id = $contractor_id;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getTaskId(): ?int
+    {
+        return $this->task_id;
+    }
+
+    /**
+     * @param int $task_id
+     */
+    public function setTaskId(int $task_id): void
+    {
+        $this->task_id = $task_id;
     }
 }
