+++ models/RespondsProblem.php
@@ -4,6 +4,7 @@ namespace app\models;
 
 use app\models\interfaces\RespondsInterface;
 use app\models\traits\SoftDeleteModelTrait;
+use Yii;
 use yii\base\ErrorException;
 use yii\db\ActiveQuery;
 use yii\helpers\FileHelper;
@@ -23,6 +24,8 @@ use yii\db\ActiveRecord;
  * @property int $date_plan                             Плановая дата интервью
  * @property string $place_interview                    Место проведения интервью
  * @property int|null $deleted_at                       Дата удаления
+ * @property int|null $contractor_id                    Идентификатор исполнителя, который опросил респондента (если null - опрос проводил проектант)
+ * @property int|null $task_id                          Идентификатор задания исполнителя, по которому исполнитель опросил респондента (если null - опрос проводил проектант)
  *
  * @property ConfirmProblem $confirm                    Подтверждение проблемы
  * @property InterviewConfirmProblem $interview         Информация о проведении интервью
@@ -98,6 +101,7 @@ class RespondsProblem extends ActiveRecord implements RespondsInterface
             [['name'], 'string', 'max' => 100],
             [['info_respond', 'email', 'place_interview'], 'string', 'max' => 255],
             ['email', 'email', 'message' => 'Неверный формат адреса электронной почты'],
+            [['contractor_id', 'task_id'], 'safe'],
         ];
     }
 
@@ -141,10 +145,11 @@ class RespondsProblem extends ActiveRecord implements RespondsInterface
 
 
     /**
+     * @param bool $isRemoveAnswers
      * @return void
      * @throws ErrorException
      */
-    private function deleteDataRespond(): void
+    private function deleteDataRespond(bool $isRemoveAnswers = true): void
     {
         $confirm = ConfirmProblem::findOne($this->getConfirmId());
         $problem = Problems::findOne($confirm->getProblemId());
@@ -158,7 +163,11 @@ class RespondsProblem extends ActiveRecord implements RespondsInterface
         }
         //Удаление ответов респондента на вопросы интервью
         if (AnswersQuestionsConfirmProblem::findAll(['respond_id' => $this->getId()])) {
-            AnswersQuestionsConfirmProblem::deleteAll(['respond_id' => $this->getId()]);
+            if ($isRemoveAnswers) {
+                AnswersQuestionsConfirmProblem::deleteAll(['respond_id' => $this->getId()]);
+            } else {
+                AnswersQuestionsConfirmProblem::updateAll(['answer' => ''], ['respond_id' => $this->getId()]);
+            }
         }
         //Удаление дирректории респондента
         $del_dir = UPLOAD.'/user-'.$user->getId().'/project-'.$project->getId().'/segments/segment-'.$segment->getId().'/problems/problem-'.$problem->getId().'/interviews/respond-'.$this->getId();
@@ -171,6 +180,43 @@ class RespondsProblem extends ActiveRecord implements RespondsInterface
             FileHelper::removeDirectory($cachePathDelete);
         }
     }
+
+
+    /**
+     * @return bool
+     */
+    public function clearData(): bool
+    {
+        $transaction = Yii::$app->db->beginTransaction();
+
+        try {
+            $responds = self::findAll(['confirm_id' => $this->getConfirmId()]);
+            $numberResponds = [];
+            foreach ($responds as $respond) {
+                if (preg_match('/^Респондент \d+$/', $respond->getName())) {
+                    $numberResponds[] = str_replace('Респондент ', '', $respond->getName());
+                }
+            }
+
+            $this->name = 'Респондент ' . (max($numberResponds) + 1);
+            $this->info_respond = '';
+            $this->email = '';
+            $this->date_plan = null;
+            $this->place_interview = '';
+            $this->contractor_id = null;
+            $this->task_id = null;
+            $this->save();
+            $this->deleteDataRespond(false);
+
+            $transaction->commit();
+            return true;
+
+        } catch (\Exception $exception) {
+            $transaction->rollBack();
+            return false;
+        }
+    }
+
 
     /**
      * @return int
@@ -293,6 +339,38 @@ class RespondsProblem extends ActiveRecord implements RespondsInterface
     public function setDeletedAt(int $deleted_at): void
     {
         $this->deleted_at = $deleted_at;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getContractorId(): ?int
+    {
+        return $this->contractor_id;
+    }
+
+    /**
+     * @param int $contractor_id
+     */
+    public function setContractorId(int $contractor_id): void
+    {
+        $this->contractor_id = $contractor_id;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getTaskId(): ?int
+    {
+        return $this->task_id;
+    }
+
+    /**
+     * @param int $task_id
+     */
+    public function setTaskId(int $task_id): void
+    {
+        $this->task_id = $task_id;
     }
 
 }
